DPI Scaling: Fix integer division issue
* Convert dpi to a qreal
  (Prevents uiScale from being clamped to an integer value)

Open-WebOS-DCO-1.0-Signed-Off-By: Josh Palmer <jpalmerwatkins@me.com>

+++ Src/base/settings/Settings.h
@@ -168,7 +168,7 @@ public:
 	std::string         launcherScrim;
 	std::string         firstCardLaunch;
 	bool                atlasEnabled;
-	int                 dpi;
+	qreal               dpi;
 	qreal               uiScale;
 	qreal               imeScale;
 	qreal               webAppScale;
